Se crea el módulo de cámara
